feature: MVP backend
Se ha hecho lo siguiente: Modularizado la API, Mejorado la documentación de la API, Corregido errores varios. Falta: Testear cada endpoint
