+++ app-back/Dockerfile
@@ -3,4 +3,4 @@ WORKDIR /app-back
 RUN pip install fastapi uvicorn mysql-connector-python sqlalchemy pydantic
 COPY . .
 EXPOSE 8000
-CMD uvicorn backend-service:app --host 0.0.0.0 --port 8000
+CMD uvicorn backend-service:app --host 0.0.0.0 --port 8000 --reload
